Documentation, and refactoring of file locations
